Update setup.sh
When running this script, gcloud failed asking for the --zone variable, so I added it to the script in case someone else runs into this issue. It also asks for the password each time it does an ssh to the controller.  I am wondering if we should try to combine this into one step so we can minimize user intervention.